requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
pydicom
opencv-python
pillow
matplotlib
tqdm
scipy
